Move site & theme configs into config dir
